fixed out of bounds, which path, value of salaries

+++ src/phase1/Cards/F1Race.java
@@ -10,7 +10,7 @@ public class F1Race extends BlueCard {
     @Override
     public void activate(Player p, Player[] otherPlayers) {
         Player playerMatched = this.othersMatched(otherPlayers);
-        double amount = p.getSalaryCard().getSalary() * 0.10;
+        double amount = (int) p.getSalaryCard().getSalary() * 0.10;
         if (this.careerMatches(p.getCareer())) {
             p.updateCash(amount);
         } else if (playerMatched != null) {
